# Agent restructuring