--- a/build.gradle
+++ b/build.gradle
@@ -15,6 +15,7 @@ buildscript {
         classpath "com.netflix.nebula:gradle-aggregate-javadocs-plugin:3.0.1"
         classpath "com.google.protobuf:protobuf-gradle-plugin:0.8.5"
         classpath "io.franzbecker:gradle-lombok:1.14"
+        classpath "com.moowork.gradle:gradle-node-plugin:1.2.0"
     }
 }