--- a/genie-agent-app/build.gradle
+++ b/genie-agent-app/build.gradle
@@ -1,0 +1,41 @@
+apply plugin: "org.springframework.boot"
+
+license {
+    exclude "*.yml"
+    exclude "*.xml"
+    exclude "*.txt"
+}
+
+dependencies {
+    /*******************************
+     * Compile Dependencies
+     *******************************/
+
+    compile project(":genie-agent")
+
+    /*******************************
+     * Provided Dependencies
+     *******************************/
+
+    /*******************************
+     * Runtime Dependencies
+     *******************************/
+
+    /*******************************
+     * Test Dependencies
+     *******************************/
+
+    testCompile(project(":genie-test"))
+}
+
+springBoot {
+    mainClassName = "com.netflix.genie.GenieAgentApplication"
+}
+
+def genieVersion = project.version.toString()
+
+jar {
+    manifest {
+        attributes("Implementation-Version": genieVersion)
+    }
+}