--- a/genie-agent/build.gradle
+++ b/genie-agent/build.gradle
@@ -6,6 +6,7 @@ license {
     exclude "*.yml"
     exclude "*.xml"
     exclude "*.txt"
+    exclude "META-INF/spring.factories"
 }
 
 dependencies {