--- a/genie-agent/src/main/java/com/netflix/genie/agent/cli/CliAutoConfiguration.java
+++ b/genie-agent/src/main/java/com/netflix/genie/agent/cli/CliAutoConfiguration.java
@@ -24,6 +24,7 @@ import com.netflix.genie.agent.execution.services.AgentJobService;
 import com.netflix.genie.agent.execution.services.DownloadService;
 import com.netflix.genie.agent.execution.services.KillService;
 import com.netflix.genie.agent.execution.statemachine.JobExecutionStateMachine;
+import com.netflix.genie.common.internal.dto.v4.converters.JobServiceProtoConverter;
 import com.netflix.genie.proto.PingServiceGrpc;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -42,10 +43,16 @@ import javax.validation.Validator;
 @Configuration
 public class CliAutoConfiguration {
 
-//    @Bean
-//    public JobServiceProtoConverter jobServiceProtoConverter() {
-//        return new JobServiceProtoConverter();
-//    }
+    /**
+     * Provide a bean for proto conversion of job service objects if one isn't already defined.
+     *
+     * @return A {@link JobServiceProtoConverter} instance.
+     */
+    @Bean
+    @ConditionalOnMissingBean(JobServiceProtoConverter.class)
+    public JobServiceProtoConverter jobServiceProtoConverter() {
+        return new JobServiceProtoConverter();
+    }
 
     /**
      * Provide a bean for cache command line arguments if none was created already.