Restructure agent to improve ability to modify behavior internally and improve performance by tuning Spring Boot based on recommendations mostly found [here](http://presos.dsyer.com/decks/how-fast-is-spring.html) and [here](https://github.com/dsyer/spring-boot-startup-bench)