@tgianos 2018-10-11
@mprimi @sumitnetflix If you want take a final look over this. I don't think I'll merge it before I have the internal wrapper ready to go. I don't want to break testing of the agent if we don't have a replacement ready but that shouldn't take too long