@tgianos 2018-10-11
This PR drive startup times down. Benchmarks run using https://github.com/dsyer/spring-boot-startup-bench/tree/master/launcher. Also improves the ability to modify agent code in wrappers since business logic is in thin jar with auto configs.

Before
```
# Run complete. Total time: 00:02:01

Benchmark               Mode  Cnt  Score   Error  Units
MainBenchmark.jar       avgt   10  3.559 ± 0.317   s/op
MainBenchmark.launcher  avgt   10  3.697 ± 0.171   s/op
MainBenchmark.main      avgt   10  2.523 ± 0.260   s/op
```

After:
```
# Run complete. Total time: 00:01:49

Benchmark               Mode  Cnt  Score   Error  Units
MainBenchmark.jar       avgt   10  3.288 ± 0.306   s/op
MainBenchmark.launcher  avgt   10  3.261 ± 0.354   s/op
MainBenchmark.main      avgt   10  2.281 ± 0.209   s/op
```